[hlm-1234]: Move app to separate file
